Latest and Greatest

+++ FriendlyForms/RestService/UserRestService.cs
@@ -1,7 +1,8 @@
-﻿using System.Collections.Generic;
+﻿using System.Net;
 using BusinessLogic.Contracts;
 using Models;
 using ServiceStack.Common;
+using ServiceStack.Common.Web;
 using ServiceStack.ServiceHost;
 using ServiceStack.ServiceInterface;
 
@@ -9,46 +10,32 @@ namespace FriendlyForms.RestService
 {
     public class UserRestService
     {
-        //REST Resource DTO
-        [Route("/Users")]
-        [Route("/Users/{Ids}")]
-        public class UserListDto : IReturn<List<UserDto>>
-        {
-            public long[] Ids { get; set; }
-
-            public UserListDto(params long[] ids)
-            {
-                Ids = ids;
-            }
-        }
-
+        ////REST Resource DTO
         [Route("/Users", "POST")]
         [Route("/Users/", "PUT")]
-        [Route("/Users/{Id}", "GET")]
         [Route("/Users", "GET")]
+        [Route("/Users/{Id}")]
         public class UserDto : IReturn<UserDto>
         {
             public long Id { get; set; }
+            public long[] Ids { get; set; }
             public int UserAuthId { get; set; }
             public bool Verified { get; set; }
             public int? LawFirmId { get; set; }
             public string Position { get; set; }
         }
-
+        
         public class UsersService : Service
         {
             public IUserService UserService { get; set; } //Injected by IOC
 
             public object Get(UserDto request)
             {
-                var user = UserService.Get(request.Id);
-                return user;
-            }
-
-            public object Get(UserListDto request)
-            {
-                //TODO Do something more interested.  Add query possibly 
-                return UserService.GetFiltered(t => t.Id != 0);
+                if(request.Ids != null && request.Ids.Length > 0)
+                    return UserService.Get(request.Ids);
+                if(request.Id > 0)
+                    return UserService.Get(request.Id);
+                throw new HttpError(HttpStatusCode.BadRequest, "Invalid arguments supplied.");
             }
 
             public object Post(UserDto request)
@@ -65,15 +52,19 @@ namespace FriendlyForms.RestService
                 return userEntity;
             }
 
-            public void Delete(UserListDto request)
-            {
-                UserService.DeleteAll(request.Ids);
-            }
-
             public void Delete(UserDto request)
             {
-                var userEntity = request.TranslateTo<User>();
-                UserService.Delete(userEntity);
+                if (request.Ids != null && request.Ids.Length > 0)
+                {
+                    UserService.DeleteAll(request.Ids);
+                    return;
+                }                    
+                if (request.Id > 0)
+                {
+                    UserService.Delete(request.Id);
+                    return;
+                }                    
+                throw new HttpError(HttpStatusCode.BadRequest, "Invalid arguments supplied.");
             }
         }
